source element's src === link's href
I always intended for the source element's src and the link element's
href to be exactly the same. What happened was that I was setting the
src to equal the link element's href property, which ends up being the
resolved version of the location. This is all fine and good on a live
web page but, when processing HTML with node it was causing relative
hrefs to be resolved to the local filesystem and then written as the
value of the src property. This wouldn't work out well at all. So I
changed this function to set the src as the returned result of the
getAttribute('href') call on the link element. This gives me the
unaltered version of the href and HTML may be processed server side
without breaking.